refactor(automations): extract cron detail panel
